some more changes trying to fix issues
